# api/docs/swagger.yaml
definitions:
  dtos.NewQuestionDto:
    properties:
      anonymous:
        type: boolean
      text:
        type: string
    required:
    - text
    type: object
  dtos.QuestionDto:
    properties:
      anonymous:
        type: boolean
      answered:
        type: boolean
      creator:
        type: string
      id:
        type: string
      owned:
        type: boolean
      text:
        type: string
      votes:
        type: integer
    type: object
info:
  contact: {}
  description: A voting tool API in Go using Gin framework.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Voting tool api
  version: "1.0"
paths:
  /question/answer/{id}:
    put:
      description: Answers a question of the current session
      parameters:
      - description: Id of question to answer
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
        "404":
          description: Not Found
          schema:
            type: string
      security:
      - JWT: []
      summary: Answers a question
      tags:
      - Question
  /question/new:
    post:
      description: Adds a new question to the current session
      parameters:
      - description: Question JSON
        in: body
        name: question
        required: true
        schema:
          $ref: '#/definitions/dtos.NewQuestionDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
      security:
      - JWT: []
      summary: Adds a new question
      tags:
      - Question
  /question/session:
    get:
      description: Gets the questions of the current session
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dtos.QuestionDto'
            type: array
        "401":
          description: Unauthorized
      security:
      - JWT: []
      summary: Gets the questions of the current session
      tags:
      - Question
  /question/session/start:
    post:
      description: Starts a new questions session
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
      security:
      - JWT: []
      summary: Starts a new questions session
      tags:
      - Question
  /question/session/stop:
    post:
      description: Stops the current questions session
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
      security:
      - JWT: []
      summary: Stops the current questions session
      tags:
      - Question
  /question/upvote/{id}:
    put:
      description: Upvotes a question of the current session
      parameters:
      - description: Id of question to upvote
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
        "404":
          description: Not Found
          schema:
            type: string
      security:
      - JWT: []
      summary: Upvotes a question
      tags:
      - Question
securityDefinitions:
  JWT:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"